requirements.txt
asgiref==3.3.1
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-imagekit==4.0.2
djangorestframework==3.12.4
Markdown==3.3.4
pilkit==2.0
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
